refactor(codemod): move `QueryClient` related parts and testfixutes to `useQuery` codemod
The plan is to combine all existing codemods into a single one.
