+++ codemods/v4/query_client_methods/__testfixtures__/named-import.output.tsx
@@ -1,52 +1,5 @@
 import * as React from 'react'
-import { useQueryClient as useRenamedQueryClient } from 'react-query'
 
 export const Examples = () => {
-  // Instantiated hook call.
-  const queryClient = useRenamedQueryClient()
-  queryClient.getMutationDefaults(['todos'])
-  queryClient.getQueriesData(['todos'])
-  queryClient.getQueryData(['todos'])
-  queryClient.getQueryDefaults(['todos'])
-  queryClient.getQueryState(['todos'])
-  queryClient.isFetching(['todos'])
-  queryClient.setMutationDefaults(['todos'], { mutationFn: async () => null })
-  queryClient.setQueriesData(['todos'], () => null)
-  queryClient.setQueryData(['todos'], () => null)
-  queryClient.setQueryDefaults(['todos'], { queryFn: async () => null })
-  queryClient.cancelQueries(['todos'])
-  queryClient.fetchInfiniteQuery(['todos'])
-  queryClient.fetchQuery(['todos'])
-  queryClient.invalidateQueries(['todos'])
-  queryClient.prefetchInfiniteQuery(['todos'])
-  queryClient.prefetchQuery(['todos'])
-  queryClient.refetchQueries(['todos'])
-  queryClient.removeQueries(['todos'])
-  queryClient.resetQueries(['todos'])
-  // Direct hook call.
-  useRenamedQueryClient().getMutationDefaults(['todos'])
-  useRenamedQueryClient().getQueriesData(['todos'])
-  useRenamedQueryClient().getQueryData(['todos'])
-  useRenamedQueryClient().getQueryDefaults(['todos'])
-  useRenamedQueryClient().getQueryState(['todos'])
-  useRenamedQueryClient().isFetching(['todos'])
-  useRenamedQueryClient().setMutationDefaults(['todos'], {
-    mutationFn: async () => null,
-  })
-  useRenamedQueryClient().setQueriesData(['todos'], () => null)
-  useRenamedQueryClient().setQueryData(['todos'], () => null)
-  useRenamedQueryClient().setQueryDefaults(['todos'], {
-    queryFn: async () => null,
-  })
-  useRenamedQueryClient().cancelQueries(['todos'])
-  useRenamedQueryClient().fetchInfiniteQuery(['todos'])
-  useRenamedQueryClient().fetchQuery(['todos'])
-  useRenamedQueryClient().invalidateQueries(['todos'])
-  useRenamedQueryClient().prefetchInfiniteQuery(['todos'])
-  useRenamedQueryClient().prefetchQuery(['todos'])
-  useRenamedQueryClient().refetchQueries(['todos'])
-  useRenamedQueryClient().removeQueries(['todos'])
-  useRenamedQueryClient().resetQueries(['todos'])
-
   return <div>Example Component</div>
 }
