+++ codemods/v4/use_query/__testfixtures__/parameter-is-identifier.output.tsx
@@ -1,10 +1,16 @@
 import * as React from 'react'
-import { useMutation, useQuery } from 'react-query'
+import { useMutation, useQuery, useQueryClient } from 'react-query'
 
 export const ExampleWithStringLiteralKey = () => {
   const stringLiteralKey = 'todos'
   useQuery([stringLiteralKey])
   useMutation([stringLiteralKey])
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries([stringLiteralKey])
+  // --- Direct hook call.
+  useQueryClient().cancelQueries([stringLiteralKey])
 
   return <div>Example Component</div>
 }
@@ -19,6 +25,12 @@ export const ExampleWithArrayExpressionKey = () => {
   const arrayExpressionKey = ['todos']
   useQuery(arrayExpressionKey)
   useMutation(arrayExpressionKey)
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries(queryKey2)
+  // --- Direct hook call.
+  useQueryClient().cancelQueries(queryKey2)
 
   return <div>Example Component</div>
 }
@@ -26,6 +38,12 @@ export const ExampleWithArrayExpressionKey = () => {
 export const ExampleWithUnknownKey = () => {
   useQuery(unknownQueryKey)
   useMutation(unknownQueryKey)
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries(unknownQueryKey)
+  // --- Direct hook call.
+  useQueryClient().cancelQueries(unknownQueryKey)
 
   return <div>Example Component</div>
 }
