+++ codemods/v4/use_query/__testfixtures__/parameter-is-object-expression.output.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { useMutation, useQuery } from 'react-query'
+import { useMutation, useQuery, useQueryClient } from 'react-query'
 
 const options = {}
 
@@ -10,6 +10,18 @@ export const ExampleWithStringLiteral = () => {
   }, options)
   useMutation({
     mutationKey: ['todos'],
+    exact: true
+  }, options)
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries({
+    queryKey: ['todos'],
+    exact: true
+  }, options)
+  // --- Direct hook call.
+  useQueryClient().cancelQueries({
+    queryKey: ['todos'],
     exact: true
   }, options)
 
@@ -26,6 +38,18 @@ export const ExampleWithStringLiteralIdentifier = () => {
     mutationKey: [stringLiteralKey],
     exact: true
   }, options)
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries({
+    queryKey: [stringLiteralKey],
+    exact: true
+  }, options)
+  // --- Direct hook call.
+  useQueryClient().cancelQueries({
+    queryKey: [stringLiteralKey],
+    exact: true
+  }, options)
 
   return <div>Example Component</div>
 }
@@ -37,6 +61,18 @@ export const ExampleWithTemplateLiteral = () => {
   }, options)
   useMutation({
     mutationKey: [`todos`],
+    exact: true
+  }, options)
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries({
+    queryKey: [`todos`],
+    exact: true
+  }, options)
+  // --- Direct hook call.
+  useQueryClient().cancelQueries({
+    queryKey: [`todos`],
     exact: true
   }, options)
 
@@ -53,6 +89,18 @@ export const ExampleWithTemplateLiteralIdentifier = () => {
     mutationKey: [templateLiteralKey],
     exact: true
   }, options)
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries({
+    queryKey: [templateLiteralKey],
+    exact: true
+  }, options)
+  // --- Direct hook call.
+  useQueryClient().cancelQueries({
+    queryKey: [templateLiteralKey],
+    exact: true
+  }, options)
 
   return <div>Example Component</div>
 }
@@ -64,6 +112,18 @@ export const ExampleWithArrayExpression = () => {
   }, options)
   useMutation({
     mutationKey: ['todos'],
+    exact: true
+  }, options)
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries({
+    queryKey: ['todos'],
+    exact: true
+  }, options)
+  // --- Direct hook call.
+  useQueryClient().cancelQueries({
+    queryKey: ['todos'],
     exact: true
   }, options)
 
@@ -80,6 +140,18 @@ export const ExampleWithArrayExpressionIdentifier = () => {
     mutationKey: arrayExpressionKey,
     exact: true
   }, options)
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries({
+    queryKey: arrayExpressionKey,
+    exact: true
+  }, options)
+  // --- Direct hook call.
+  useQueryClient().cancelQueries({
+    queryKey: arrayExpressionKey,
+    exact: true
+  }, options)
 
   return <div>Example Component</div>
 }
@@ -89,6 +161,12 @@ export const ExampleWithUnknownIdentifier1 = () => {
   const createdKey1 = createKey(1)
   useQuery({ queryKey: createdKey1, exact: true }, options)
   useMutation({ mutationKey: createdKey1, exact: true }, options)
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries({ queryKey: createdKey1, exact: true }, options)
+  // --- Direct hook call.
+  useQueryClient().cancelQueries({ queryKey: createdKey1, exact: true }, options)
 
   return <div>Example Component</div>
 }
@@ -97,6 +175,12 @@ export const ExampleWithUnknownIdentifier2 = () => {
   const createdKey2 = createKey()
   useQuery({ queryKey: createdKey2, exact: true }, options)
   useMutation({ mutationKey: createdKey2, exact: true }, options)
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries({ queryKey: createdKey2, exact: true }, options)
+  // --- Direct hook call.
+  useQueryClient().cancelQueries({ queryKey: createdKey2, exact: true }, options)
 
   return <div>Example Component</div>
 }
@@ -104,6 +188,12 @@ export const ExampleWithUnknownIdentifier2 = () => {
 export const ExampleWithUnknownIdentifier3 = () => {
   useQuery({ queryKey: unknownQueryKey, exact: true }, options)
   useMutation({ mutationKey: unknownQueryKey, exact: true }, options)
+  // QueryClient methods
+  // --- Instantiated hook call.
+  const queryClient = useQueryClient()
+  queryClient.cancelQueries({ queryKey: unknownQueryKey, exact: true }, options)
+  // --- Direct hook call.
+  useQueryClient().cancelQueries({ queryKey: unknownQueryKey, exact: true }, options)
 
   return <div>Example Component</div>
 }
